+++ www/commit-hash.js
@@ -1,1 +1,1 @@
-window.__COMMIT__ = 'f819d0875e326256357dafd90d26155fb6c87fac';
+window.__COMMIT__ = '369912108157dda9eb6b6eb5a7788c0869eac925';